Redefine transform between KinovaAngles and kinova_msgs::JointAngles
KinovaAngles are in degrees, and they are used in SDK for joint control. While
kinova_msgs::JointAngles are in radians, and they are used for display URDF
model in Rviz, and ROS tf topics.
They were different in both direction and initial value. Now they are directly
linked by a radio of 180/M_PI, with same direction and initial value.
Offset for joint 6 "j6o" is not needed for kinova_msgs. Instead it could be
parameter in URDF file, and its value is dertermined based on robot type.