Correct mico URDF to use the new kinova_msgs::JointAngles
kinova_msgs::JointAngles are modified to be coincide with .../out/joint_values
topic, which is also the widely used convention for SDK and JacoSoft, etc. The
URDF model (espeically the parameter "origin_rpy") should be redefined to
display robot properly in Rviz with this new convenient/uniform
kinova_msgs::JointAngles.