Create new message for KinovaPose
KinovaPose use EulerAngles to express robot's orientation, it is also the format
used in the SDK and JacoSoft for Kinova robots. This ROS message will be
used in both control robot Cartesian pose and reading Cartesian pose. It is a
counterpart as geometry_msg::Pose which use Quaternions for orientation.