Cartesian position control precision test
Based on various position test, the performance of API setCartesianPosition is
evaluated. In the test, a goal position is sent and final position is measured
after robot stops motion.  The result is: robot may consider it reach the goal
and stop motion when distance is less than 0.01meter or Euler angle difference
in 5degrees.

To be noted, the test didn't cover all workspace. The test is aiming to gvie a
general idea on the position control.