Fix arm drop when sending joint/cartesian command sent
The drop is related with the gravity. When an actuator bear more gravity
(horizontal), it drops more. While in 180 position, no drop for all actuators.
This happens only in the relative motion for joint/cartesian motion. The problem
can be solved by reading AngularCommand or CartesianCommand as current position,
rather than reading getAngularPose or getCartesianPose